Fix warning -Wsign-compare (#18)

+++ src/PlayState.cpp
@@ -36,7 +36,7 @@ void PlayState::update()
         object->update();
     
     SDLGameObject* player = dynamic_cast<SDLGameObject*>(gameObjects[0]); 
-    for (int i = 1; i <= gameObjects.size(); i++)
+    for (long unsigned int i = 1; i <= gameObjects.size(); i++)
     {
         if (checkCollision(player, dynamic_cast<SDLGameObject*>(gameObjects[i])))
             Game::Instance()->getStateMachine()->pushState(new GameOverState());
